Resolve symlink for checking whitelist
On POSIX systems the whitelist now checks both against the link
name and against the real path. Before it checked only the link
name, when the application encountered the link.

Thank you to Vladimir Chapaikin  for the bug report.

Fixes
https://bugs.launchpad.net/bleachbit/+bug/1455959

There is a speed penalty of about 50%. The benchmark went
up from 1.6 seconds to 2.4 seconds.

R code:
t.test(c(2.34,2.36, 2.46, 2.38), c(1.52, 1.53, 1.56, 1.62))

But it was practically insignificant because this was a cost of
0.8 seconds over 20,000 files.